pyproject.toml
[project]
name = "xai-demo-suite"
version = "0.1.0"
description = "Curated explainable AI demos for vision, shortcut learning, and PatchCore anomaly detection."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "John Goodacre" }
]
dependencies = [
  "numpy>=1.26",
  "pillow>=10.0",
  "pyyaml>=6.0",
]

[dependency-groups]
dev = [
  "mypy>=1.10",
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "ruff>=0.5",
]
notebooks = [
  "jupyter>=1.0",
  "jupytext>=1.16",
  "nbmake>=1.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "C4", "SIM", "RUF"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_configs = true
mypy_path = "src"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.coverage.run]
source = ["xai_demo_suite"]
branch = true
